examples/function_samples/grpc_echo_sample/client/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
Pillow==10.0.0
gradio==3.9.1
httpx==0.24.1